backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
google-auth==2.30.0
qrcode==7.4.2
Pillow==10.4.0
